swagger_server/controllers/contenidos_controller.py
import connexion
import six

from swagger_server.models.actor import Actor  # noqa: E501
from swagger_server.models.contenido import Contenido  # noqa: E501
from swagger_server.models.contenidos_body import ContenidosBody  # noqa: E501
from swagger_server.models.contenidos_body1 import ContenidosBody1  # noqa: E501
from swagger_server.models.episodio import Episodio  # noqa: E501
from swagger_server.models.genero import Genero  # noqa: E501
from swagger_server import util


def contenidos_filtrar_get(titulo=None, nombre_actor=None, nombre_genero=None):  # noqa: E501
    """Filtrar contenidos por nombre, actor y género

    Devuelve una lista de contenidos que coincidan con el nombre, actor o género especificado. # noqa: E501

    :param titulo: Nombre o título parcial del contenido a buscar
    :type titulo: str
    :param nombre_actor: Nombre del actor para filtrar los contenidos asociados a dicho actor
    :type nombre_actor: str
    :param nombre_genero: Nombre del género para filtrar los contenidos asociados a dicho género
    :type nombre_genero: str

    :rtype: List[Contenido]
    """
    return 'do some magic!'


def contenidos_get():  # noqa: E501
    """Obtener todos los contenidos

     # noqa: E501


    :rtype: List[Contenido]
    """
    return 'do some magic!'


def contenidos_id_contenido_actores_get(id_contenido):  # noqa: E501
    """Obtener todos los actores asociados a un contenido

    Devuelve una lista de actores asociados al contenido especificado por su ID. # noqa: E501

    :param id_contenido: ID del contenido para el cual se desean obtener los actores asociados
    :type id_contenido: int

    :rtype: List[Actor]
    """
    return 'do some magic!'


def contenidos_id_contenido_delete(id_contenido):  # noqa: E501
    """Eliminar un contenido

     # noqa: E501

    :param id_contenido: 
    :type id_contenido: int

    :rtype: None
    """
    return 'do some magic!'


def contenidos_id_contenido_episodios_get(id_contenido):  # noqa: E501
    """Obtener todos los episodios asociados a un contenido

    Devuelve una lista de episodios asociados al contenido especificado por su ID. # noqa: E501

    :param id_contenido: ID del contenido para el cual se desean obtener los episodios asociados
    :type id_contenido: int

    :rtype: List[Episodio]
    """
    return 'do some magic!'


def contenidos_id_contenido_generos_get(id_contenido):  # noqa: E501
    """Obtener todos los géneros asociados a un contenido

    Devuelve una lista de géneros asociados al contenido especificado por su ID. # noqa: E501

    :param id_contenido: ID del contenido para el cual se desean obtener los géneros asociados
    :type id_contenido: int

    :rtype: List[Genero]
    """
    return 'do some magic!'


def contenidos_id_contenido_get(id_contenido):  # noqa: E501
    """Obtener un contenido por ID

     # noqa: E501

    :param id_contenido: 
    :type id_contenido: int

    :rtype: Contenido
    """
    return 'do some magic!'


def contenidos_post(body):  # noqa: E501
    """Crear un nuevo contenido con actores y géneros

    Crea un contenido nuevo y asocia múltiples actores y géneros. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Contenido
    """
    if connexion.request.is_json:
        body = ContenidosBody1.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def contenidos_put(body):  # noqa: E501
    """Actualizar un contenido con actores y géneros asociados

    Actualiza los datos de un contenido existente, incluyendo sus relaciones con actores y géneros. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Contenido
    """
    if connexion.request.is_json:
        body = ContenidosBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'
